% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{choose_cpp}
\alias{choose_cpp}
\title{n choose k helper for combn_cpp}
\usage{
choose_cpp(n, k)
}
\arguments{
\item{n}{integer number of elements to choose from}

\item{k}{integer number of elements to choose}
}
\value{
integer number of ways to choose k items out of n
}
\description{
n choose k helper for combn_cpp
}
\keyword{internal}
