% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combn_cpp}
\alias{combn_cpp}
\title{helper to generate combinations of x k at a time (cpp implementation of combn)}
\usage{
combn_cpp(x, K)
}
\arguments{
\item{x}{integer vector of elements to combine}

\item{K}{integer order of combinations}
}
\value{
a matrix of k-wise combinations of elements in x
}
\description{
helper to generate combinations of x k at a time (cpp implementation of combn)
}
\keyword{internal}
