% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gen_block_matrix}
\alias{gen_block_matrix}
\title{Generate block matrix of link probabilities}
\usage{
gen_block_matrix(p_edge_within, p_edge_between)
}
\arguments{
\item{p_edge_within}{named list of numeric vectors giving probability of link between in-group members and out-group members for each of groups}

\item{p_edge_between}{named list of numeric values giving probability of link between in- and out-group member for each of groups}
}
\value{

}
\description{
Returns full block matrix of link probabilities given the group memberships based on probabilities of links within each of the groups and between in- and out-group members
}
\examples{
\dontrun{
gen_block_matrix(p_edge_within = list(known = c(0.05, 0.05), hidden = c(0.05, 0.9)),
                 p_edge_between = list(known = 0.05, hidden = 0.01))
}

}
