% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gen_marginal_types}
\alias{gen_marginal_types}
\title{Generate vector of type labels}
\usage{
gen_marginal_types(k)
}
\arguments{
\item{k}{number of groups}
}
\value{

}
\description{
Generate vector of type labels
}
\examples{
gen_marginal_types(3)

}
