% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_range}
\alias{gen_range}
\title{helper to generate range of consecutive integers}
\usage{
gen_range(from, to)
}
\arguments{
\item{from}{integer first integer}

\item{to}{last integer}
}
\value{
integer vector of consecutive integers between from and to inclusive
}
\description{
helper to generate range of consecutive integers
}
\keyword{internal}
