% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimands.R
\name{get_hpop_estimands}
\alias{get_hpop_estimands}
\title{Get estimands for several studies}
\usage{
get_hpop_estimands(data)
}
\arguments{
\item{data}{pass-through data frame of several study populations}
}
\value{

}
\description{
Get estimands for several studies
}
