% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_multistudy.R
\name{get_hpop_sample}
\alias{get_hpop_sample}
\title{Draw samples from multiple study populations according to proposed strategy}
\usage{
get_hpop_sample(
  data,
  sampling_variable = "rds",
  drop_nonsampled = FALSE,
  sampling_args,
  sampling_strategy = c("sample_rds", "sample_prop")
)
}
\arguments{
\item{data}{pass-through populations data frame}

\item{sampling_variable}{character string that is used as prefix for all variables generated by RDS sampling (sample identifier, recruiter ID, wave, time of show-up) in each population}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{sampling_args}{named list of named lists of arguments to \code{sample_rds()} (the set of arguments can be excessive)}

\item{sampling_strategy}{character string giving name of the sampling strategy handler}
}
\value{

}
\description{
Sampling handler for drawing samples with given characteristics from multiple study populations. The sampling characteristics can vary across studies
}
