% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_estimands.R
\name{get_meta_estimands}
\alias{get_meta_estimands}
\title{Get estimands for several studies}
\usage{
get_meta_estimands(data)
}
\arguments{
\item{data}{pass-through data frame of several study populations}
}
\value{
Data frame of estimands for multiple studies
}
\description{
Get estimands for several studies
}
