% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_estimator.R
\name{get_meta_estimates_old}
\alias{get_meta_estimates_old}
\title{Use Stan model to estimate bias and estimands}
\usage{
get_meta_estimates_old(
  data,
  sample_label = "meta",
  which_estimand = "hidden_size",
  benchmark = list(sample = "pps", estimator = "ht"),
  stan_handler = get_meta_stan3,
  hidden_prior = NULL,
  rel_bias_prior = list(mean = 1, se = 10),
  control_params = list(iter = 8000, chains = 8, thin = 10, seed = 872312, cores = 1)
)
}
\arguments{
\item{data}{pass-through meta population or meta sample data frame}

\item{sample_label}{name of variable storing meta analysis sampling information}

\item{which_estimand}{name of study level estimand for meta analysis}

\item{benchmark}{named list of length 2 giving benchmark sampling-estimator pair (only accepts one value across studies for now)}

\item{stan_handler}{function that takes stan_data as input and produces compilable stan model object}

\item{hidden_prior}{list of two hyperpriors, on means and standard errors of each included sampling-estimator pairs. Names of list objects should be "mean" and "se". If one number provided for a hyperprior it gets expanded to all sampling-estimator pairs}

\item{rel_bias_prior}{list of two hyperpriors, on means and standard errors of relative bias. Names of list objects should be "mean" and "se".}

\item{control_params}{list of additional parameters to pass to Stan fit function. These can include number of iterations, chains, thinning, seed and number of cores to use}
}
\value{
Data frame of meta level estimates and pertaining estimand names
}
\description{
Use Stan model to estimate bias and estimands
}
