% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_population.R
\name{get_meta_population}
\alias{get_meta_population}
\title{Simulate populations with given network structure for multiple studies with (possibly) varying designs}
\usage{
get_meta_population(pop_args)
}
\arguments{
\item{pop_args}{named list of named lists of arguments to \code{get_study_population()} (the set of arguments can be excessive)}
}
\value{

}
\description{
Simulate populations with given network structure for multiple studies with (possibly) varying designs
}
