% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_sample.R
\name{get_meta_sample}
\alias{get_meta_sample}
\title{Sample sampling-estimator strategies across studies for meta-analysis}
\usage{
get_meta_sample(
  data,
  drop_nonsampled = FALSE,
  sampling_variable = "meta",
  selection_variables = c("sample", "estimator"),
  samples_per_study = 2,
  estimator_per_sample = 3,
  force = list(sample = "pps", estimator = "ht")
)
}
\arguments{
\item{data}{pass-through meta population}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{sampling_variable}{character string that is used as prefix for all variables generated by TLS sampling. Default is 'meta'}

\item{selection_variables}{length 2 character vector of names of sample and estimator variables in population data frame (in this order). Default is 'sample' and 'estimator'}

\item{samples_per_study}{number of sampling strategies per study to draw}

\item{estimator_per_sample}{number of estimation strategies to draw per sampling strategy}

\item{force}{named list of length two giving the sample and estimator identifiers that are forced to be sampled across studies}
}
\value{

}
\description{
Sample sampling-estimator strategies across studies for meta-analysis
}
