% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_meta.R
\name{get_meta_stan3}
\alias{get_meta_stan3}
\title{Helper for meta-analysis}
\usage{
get_meta_stan3(stan_data)
}
\arguments{
\item{stan_data}{A list of data for Stan}
}
\value{
A string of Stan code
}
\description{
Helper for meta-analysis
}
