% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_multi.R
\name{get_multi_estimands}
\alias{get_multi_estimands}
\title{Calculate estimands from multiple study populations}
\usage{
get_multi_estimands(data, study_designs)
}
\arguments{
\item{data}{pass-through populations data frame}

\item{study_designs}{named list of arguments for inquiries used in each study}
}
\value{
Data frame with estimands and estimand labels
}
\description{
Handler for calculating estimand(s) from multiple study populations
}
