% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_multi.R
\name{get_multi_estimates}
\alias{get_multi_estimates}
\title{Calculate estimators from multiple study samples}
\usage{
get_multi_estimates(data, estimators_args)
}
\arguments{
\item{data}{pass-through populations data frame}

\item{estimators_args}{named list of named lists of arguments to estimation functions for each study and sampling strategy}
}
\value{
Data frame with estimands and estimand labels
}
\description{
Handler for calculating estimator(s) from multiple study samples
}
