% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_multi.R
\name{get_multi_populations}
\alias{get_multi_populations}
\title{Draw population(s) with given network structure for multiple studies}
\usage{
get_multi_populations(study_designs, cl = NULL)
}
\arguments{
\item{study_designs}{named list of arguments to \code{get_study_population()} for each study.}
}
\value{
Meta tibble data frame with columns
\itemize{
 \item{"study"}{Character study identifier}
 \item{"population"}{List of tibble data frames with whole population information}
}
}
\description{
Handler for drawing populations with given characteristics for multiple studies. The population characteristics can vary across studies
}
