% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_pop_network}
\alias{get_pop_network}
\title{Simulate population with given network structure for multiple studies with varying designs}
\usage{
get_pop_network(pop_args)
}
\arguments{
\item{pop_args}{named list of named lists of arguments to pop_network()}
}
\value{

}
\description{
Simulate population with given network structure for multiple studies with varying designs
}
