% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_populations}
\alias{get_populations}
\title{Simulate populations with given network structure for multiple studies with (possibly) varying designs}
\usage{
get_populations(pop_args)
}
\arguments{
\item{pop_args}{named list of named lists of arguments to \code{pop_network()} (the set of arguments can be excessive)}
}
\value{

}
\description{
Simulate populations with given network structure for multiple studies with (possibly) varying designs
}
