% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bootstrap.R
\name{get_rds_boot}
\alias{get_rds_boot}
\title{Prepare RDS data for Bootstrap Re-sampling Following Salganik (2006)}
\usage{
get_rds_boot(
  data,
  seed_condition = "rds_from == -999",
  in_coupon = "rds_own_coupon",
  out_coupon = "rds_coupon_",
  is_child_fn = is_child,
  trait_var = "hidden_visible_out",
  other_vars = c("hidden_visible_out", "known_visible_out", "hidden"),
  n_boot = 100
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{seed_condition}{character string containing condition to define seeds. Defaults to "rds_from == -999" that applies to simulated RDS samples}

\item{in_coupon}{character string for variable containing in coupons identifiers. In coupon is coupon used to enroll specific respondent}

\item{out_coupon}{character string for variable(s) containing out coupons identifiers. Out coupons are coupons given to respondent to enroll other study participants from their network}

\item{is_child_fn}{function used to identify direct childs of specific seed. Defaults to function that uses "-" in the coupon identifiers}

\item{trait_var}{variable used to split RDS sample for resampling. Defaults to "hidden_visible_out" that captures number of hidden population members recognized by respondent}

\item{other_vars}{other variables to preserve in bootstrap samples}

\item{n_boot}{number of bootstrap resamples}
}
\value{

}
\description{
Prepare RDS data for Bootstrap Re-sampling Following Salganik (2006)
}
\examples{

}
\references{
Dennis M. Feehan, Matthew J. Salganik. “The surveybootstrap package.” (2016). \url{https://CRAN.R-project.org/package=surveybootstrap}.
}
