% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{get_required_data}
\alias{get_required_data}
\title{Generate Data Frame With Required Data by Study}
\usage{
get_required_data(multi_study_data)
}
\arguments{
\item{multi_study_data}{tibble with two columns, study and population, generate by multi-study declaration}
}
\value{

}
\description{
Generate Data Frame With Required Data by Study
}
\examples{

}
