% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bootstrap.R
\name{get_rescaled_boot}
\alias{get_rescaled_boot}
\title{Modified Function for Rescaled Bootstrap by Rust and Rao (1996)}
\usage{
get_rescaled_boot(data, survey_design, n_boot = 10)
}
\arguments{
\item{data}{sample data frame}

\item{survey_design}{a formula describing the design of the survey}

\item{n_boot}{number of bootstrap resamples}
}
\value{

}
\description{
Modified Function for Rescaled Bootstrap by Rust and Rao (1996)
}
\examples{

}
\references{
Dennis M. Feehan, Matthew J. Salganik. “The surveybootstrap package.” (2016). \url{https://cran.r-project.org/package=surveybootstrap}.
Rust, Keith F., and J. N. K. Rao. "Variance estimation for complex surveys using replication techniques." Statistical methods in medical research 5, no. 3 (1996): 283-310.
}
