% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{get_single_study_design}
\alias{get_single_study_design}
\title{Get Parameters and Design Features of Specific Study}
\usage{
get_single_study_design(
  study = "johnjay_tanzania",
  ss = "1HwMM6JwoGALLMTpC8pQRVzaQdD2X71jpZNhfpKTnF8Y",
  sheet = "study_params_readable"
)
}
\arguments{
\item{study}{Character string giving name of the study to pull. Have to match the names provided in Google Spreadsheet}

\item{ss}{Character string giving ID of the Google spreadsheet}

\item{sheet}{Character string giving name of the sheet in the Google spreadsheet to read}
}
\description{
Get Parameters and Design Features of Specific Study
}
