% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_mse}
\alias{get_study_est_mse}
\title{Multiple Systems estimator of population size for Sparse Capture Data by Chan et al.}
\usage{
get_study_est_mse(
  data,
  capture_vars = NULL,
  capture_parse = NULL,
  sample_condition = NULL,
  method = "stepwise",
  hidden_variable = "hidden",
  label = "mse"
)
}
\arguments{
\item{data}{pass-through population data frame that contains capture indicators}

\item{capture_vars}{character vector giving names of variables with capture indicators}

\item{capture_parse}{character string giving expression to evaluation of which produces character vector giving names of variable with capture indicators. Defaults to \code{NULL}. This is useful when capture variables are stored in one column (e.g. if using TLS sampled locations for recapture indicators)}

\item{sample_condition}{character string with condition if the capture-recapture conducted on subsample of population (e.g. tls sample only)}

\item{method}{character string giving the estimation method to be passed to \code{estimatepopulation.0}. See \code{?SparseMSE::estimatepopulation.0} for more details. Defaults to \code{"stepwise"}}

\item{hidden_variable}{character string giving indicator of hidden population membership}

\item{label}{character string describing the estimator}
}
\value{
Data frame of Multiple System estimates for a single study
}
\description{
Multiple Systems estimator of population size for Sparse Capture Data by Chan et al.
}
\references{
Chan, Lax, Bernard W. Silverman, and Kyle Vincent. "Multiple systems estimation for sparse capture data: Inferential challenges when there are nonoverlapping lists." Journal of the American Statistical Association 116.535 (2021): 1297-1306.
Chan, Lax, Bernard W. Silverman, and Kyle Vincent. “The SparseMSE package.” (2022). \url{https://cran.r-project.org/web/packages/SparseMSE}.
}
