% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_multiplier}
\alias{get_study_est_multiplier}
\title{Service/Object multiplier estimator}
\usage{
get_study_est_multiplier(
  data,
  service_var = "service_use",
  total_service = sum(data$service_use[data$hidden == 1]),
  seed_condition = "rds_from == -999",
  n_boot = 100,
  parallel_boot = TRUE,
  rds_prefix = "rds",
  label = "multiplier"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{service_var}{name of variable that indicates service/object use by respondent}

\item{total_service}{numeric value that indicates number of hidden population members who used the service. Defaults to truth from simulated dataset}

\item{seed_condition}{character string containing condition to define seeds. Defaults to "rds_from == -999" that applies to simulated RDS samples}

\item{n_boot}{number of bootstrap resamples}

\item{parallel_boot}{logical, whether to compute bootstrap samples in parallel using \code{foreach} package}

\item{rds_prefix}{character prefix used for RDS sample variables}

\item{label}{character string describing the estimator}
}
\value{
Data frame of service/object multiplier population size estimates for single study
}
\description{
Service/Object multiplier estimator
}
\details{
Function currently requires variable "hidden_visible_out" to be present in the data supplied and represent the hidden population out-report
}
\references{
Salganik, Matthew J. "Variance estimation, design effects, and sample size calculations for respondent-driven sampling." Journal of Urban Health 83, no. 1 (2006): 98. \url{https://doi.org/10.1007/s11524-006-9106-x}
}
