% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_nsum}
\alias{get_study_est_nsum}
\title{NSUM estimatior}
\usage{
get_study_est_nsum(
  data,
  hidden_var = "hidden_visible_out",
  known_vars = paste0(c("known", paste0("known_", 2:10)), "_visible_out"),
  total = 1000,
  prefix = "pps",
  label = "pps_nsum",
  weight_var = "pps_weight",
  survey_design = ~pps_cluster + strata(pps_strata),
  n_boot = 1000,
  parallel_boot = FALSE
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{hidden_var}{character vector containing names of hidden groups}

\item{known_vars}{character vector containing names of known groups}

\item{total}{integer giving total size of population}

\item{prefix}{character prefix used for PPS sample variables}

\item{label}{character string describing the estimator}

\item{weight_var}{character string giving name of the sampling weights variable}

\item{survey_design}{a formula describing the design of the survey}

\item{n_boot}{integer giving number of bootstrap re-samples}

\item{parallel_boot}{logical, whether to compute bootstrap samples in parallel using \code{foreach} package. Defaults to \code{FALSE}}
}
\value{
Data frame of NSUM estimates for a single study with PPS sample
}
\description{
NSUM estimatior
}
\references{
Zheng, Tian, Matthew J. Salganik, and Andrew Gelman. "How many people do you know in prison? Using overdispersion in count data to estimate social structure in networks." Journal of the American Statistical Association 101.474 (2006): 409-423.

Laga, Ian, Le Bao, and Xiaoyue Niu. "A Correlated Network Scale-up Model: Finding the Connection Between Subpopulations." Journal of the American Statistical Association just-accepted (2023): 1-18.

Salganik, Matthew J. "Variance estimation, design effects, and sample size calculations for respondent-driven sampling." Journal of Urban Health 83, no. 1 (2006): 98. \url{https://doi.org/10.1007/s11524-006-9106-x}
}
