% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecating.R
\name{get_study_est_nsum_old}
\alias{get_study_est_nsum_old}
\title{NSUM estimatior}
\usage{
get_study_est_nsum_old(
  data,
  known,
  hidden,
  survey_design = ~pps_cluster + strata(pps_strata),
  degree_ratio = 1,
  transmission_rate = 1,
  n_boot = 1000,
  parallel_boot = FALSE,
  prefix = "pps",
  label = "nsum"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{known}{character vector containing names of known groups}

\item{hidden}{character vector containing names of hidden groups}

\item{survey_design}{a formula describing the design of the survey}

\item{degree_ratio}{numeric value between 0 and 1 representing degree ratio}

\item{transmission_rate}{numeric value between 0 and 1 representing information transmission rate}

\item{n_boot}{number of bootstrap resamples}

\item{parallel_boot}{logical, whether to compute bootstrap samples in parallel using \code{foreach} package}

\item{prefix}{character prefix used for PPS sample variables}

\item{label}{character string describing the estimator}
}
\value{
Data frame of NSUM estimates for a single study with PPS sample
}
\description{
NSUM estimatior
}
\keyword{internal}
