% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_ss}
\alias{get_study_est_ss}
\title{Sequential Sampling (SS) prevalence estimator by Gile (2011)}
\usage{
get_study_est_ss(
  data,
  sampling_var = "hidden",
  hidden_var = "target",
  n_coupons = 3,
  total = 1000,
  prefix = "rds",
  label = "ss"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sampling_var}{character string specifying name of the group from which RDS sample was drawn (associated probability of visibility should be named \code{p_visible_[sampling_var]}). Defaults to "hidden" for the simulations}

\item{hidden_var}{character string specifying names of the hidden group variable name (associated probability of visibility should be named \code{p_visible_[hidden_var]}). Defaults to "target" for the simulations}

\item{n_coupons}{The number of recruitment coupons distributed to each enrolled subject (i.e. the maximum number of recruitees for any subject). By default it is taken by the attribute or data, else the maximum recorded number of coupons.}

\item{total}{integer giving the total size of known population (denominator for prevalence)}

\item{prefix}{character prefix used for RDS sample variables}

\item{label}{character string describing the estimator}
}
\value{
Data frame of SS prevalence estimates for a single study
}
\description{
Sequential Sampling (SS) prevalence estimator by Gile (2011)
}
\references{
{Gile, Krista J. 2011 Improved Inference for Respondent-Driven Sampling Data with Application to HIV Prevalence Estimation, Journal of the American Statistical Association, 106, 135-146.}

{Gile, Krista J., Handcock, Mark S., 2010 Respondent-driven Sampling: An Assessment of Current Methodology, Sociological Methodology, 40, 285-327.}
}
