% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimands.R
\name{get_study_estimands}
\alias{get_study_estimands}
\title{Get individual study estimands using data.table}
\usage{
get_study_estimands(
  data,
  in_groups = NULL,
  out_groups = "^known(\\\\_\\\\d|\\\\d)?$",
  hidden_var = "hidden"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{in_groups}{character string containing regular expression to match group names for which prevalence is estimated within hidden group}

\item{out_groups}{character string containing regular expression to match group names within which hidden group prevalence is estimated}

\item{hidden_var}{character string containing hidden group name in the study population data frame}
}
\value{
Estimands data frame for single study
}
\description{
Get individual study estimands using data.table
}
