% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecating.R
\name{get_study_estimands_tidy}
\alias{get_study_estimands_tidy}
\title{Get individual study estimands}
\usage{
get_study_estimands_tidy(
  data,
  known_pattern = "^known(\\\\_\\\\d|\\\\d)?$",
  hidden_var = "hidden"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{known_pattern}{character string containing regular expression to match known group names in the study population dataset}

\item{hidden_var}{character string containing hidden group name in the study population data frame}
}
\value{
Estimands data frame for single study
}
\description{
Get individual study estimands
}
\keyword{internal}
