% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_meta.R
\name{get_study_populations}
\alias{get_study_populations}
\title{Simulate populations with given network structure for multiple studies with (possibly) varying designs}
\usage{
get_study_populations(pops_args)
}
\arguments{
\item{pops_args}{named list of named lists of arguments to \code{get_study_population()} (the set of arguments can be excessive)}
}
\value{
Population data frame for multiple studies
}
\description{
Simulate populations with given network structure for multiple studies with (possibly) varying designs
}
