% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_meta.R
\name{get_study_samples}
\alias{get_study_samples}
\title{Draw samples from multiple study populations according to proposed strategy}
\usage{
get_study_samples(data, samples_args)
}
\arguments{
\item{data}{pass-through populations data frame}

\item{samples_args}{named list of named lists of arguments to sampling function (the set of arguments can be excessive)}
}
\value{
Population or sample data frame for multiple studies
}
\description{
Sampling handler for drawing samples with given characteristics from multiple study populations. The sampling characteristics can vary across studies
}
