% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_vec_insert}
\alias{int_vec_insert}
\title{helper to do random access insertion for vectors}
\usage{
int_vec_insert(vec, vals, pos)
}
\arguments{
\item{vec}{integer vector to insert values into}

\item{vals}{integer vector of values to insert into vec}

\item{pos}{integer vector of index positions in vec to insert vals into}
}
\value{
integer vector with vals inserted at pos
}
\description{
helper to do random access insertion for vectors
}
\keyword{internal}
