% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/helpers_linktrace.R
\name{lt_gibbs}
\alias{lt_gibbs}
\title{Link-tracing Gibbs sampler}
\usage{
lt_gibbs(
  data,
  y_samp,
  strata,
  n_strata,
  n_waves,
  total,
  chain_samples,
  chain_burnin,
  priors,
  param_init
)

lt_gibbs(
  data,
  y_samp,
  strata,
  n_strata,
  n_waves,
  total,
  chain_samples,
  chain_burnin,
  priors,
  param_init
)
}
\arguments{
\item{data}{pass through of sammple data created in get_study_est_linktrace}

\item{y_samp}{pass through of adjecency matrix generated in get_study_est_linktrace}

\item{strata}{pass through of the strata column indicator from get_study_est_linktrace}

\item{n_strata}{pass through of n_strata variable created in get_study_est_linktrace indicating number of unique strata}

\item{n_waves}{pass through of n_waves variable created in get_study_est_linktrace indicating number of rds waves}

\item{total}{pass through of total variable from get_study_est_linktrace}

\item{chain_samples}{pass through of chain_samples variable from get_study_est_linktrace indicating number of samples for gibbs sampler}

\item{chain_burnin}{pass through of chain_burnin variable from get_study_est_linktrace indicating number of burnin samples for gibbs sampler}

\item{priors}{pass through of priors from get_study_est_linktrace}

\item{param_init}{pass through of initial paramters list for population size, lambda and beta from get_study_est_linktrace}
}
\value{
integer vector of samples from the n distribution

list of gibbs sampler estimates for population size, lambda and beta
}
\description{
Link-tracing Gibbs sampler

Link-tracing gibbs sampler
}
\keyword{internal}
