% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lt_gibbs_cpp}
\alias{lt_gibbs_cpp}
\title{Link-tracing Gibbs sampler}
\usage{
lt_gibbs_cpp(
  data,
  y_samp,
  strata,
  n_strata,
  n_waves,
  total,
  chain_samples,
  priors,
  param_init
)
}
\arguments{
\item{data}{DataFrame pass through of sammple data created in get_study_est_linktrace}

\item{y_samp}{matrix pass through of adjecency matrix generated in get_study_est_linktrace}

\item{strata}{integer stratum id of each unit}

\item{n_strata}{integer number of unique strata}

\item{n_waves}{integer number of sampling waves}

\item{total}{integer total size of the population}

\item{chain_samples}{integer number of samples per MCMC chain}

\item{priors}{List of priors for n, lambda and beta}

\item{param_init}{List of initial values for n,lambda and beta}
}
\value{
integer vector of samples from the n distribution
}
\description{
Link-tracing Gibbs sampler
}
\keyword{internal}
