% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_to_mat_insert}
\alias{mat_to_mat_insert}
\title{helper to do random access insertion of matrix into another matrix}
\usage{
mat_to_mat_insert(old_m, new_m, new_rows, new_cols, old_rows, old_cols)
}
\arguments{
\item{old_m}{matrix to take values from}

\item{new_m}{matrix to insert values into}

\item{new_rows}{integer vector of row indeces of new_m to insert values into}

\item{new_cols}{integer vector of column indeces of new_m to insert values into}

\item{old_rows}{integer vector of row indeces of old_m to take values from}

\item{old_cols}{integer vector of column indeces of old_m to take values from}
}
\value{
matrix new_m with values from old_m in specified positions
}
\description{
helper to do random access insertion of matrix into another matrix
}
\keyword{internal}
