% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_elements}
\alias{move_elements}
\title{cpp helper to move vector elements to new indices}
\usage{
move_elements(x, old_index, new_index)
}
\arguments{
\item{x}{integer vector of elements to be shuffled}

\item{old_index}{integer original index of element to be moved}

\item{new_index}{integer index element should be moved to}
}
\value{
shuffled vector
}
\description{
cpp helper to move vector elements to new indices
}
\keyword{internal}
