% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population_network}
\alias{population_network}
\title{Simulate signle population with given network structure}
\usage{
population_network(
  N,
  K,
  prev_K,
  rho_K,
  p_edge_within,
  p_edge_between,
  p_visibility,
  p_service,
  directed = FALSE
)
}
\arguments{
\item{N}{number of units in population}

\item{K}{number of groups}

\item{prev_K}{numeric vector of prevalence for each group with last group being hidden population}

\item{rho_K}{numeric vector of correlations in group memberships}

\item{p_edge_within}{named list of numeric vectors giving probability of link between in-group members and out-group members for each of groups}

\item{p_edge_between}{named list of numeric values giving probability of link between in- and out-group member for each of groups}

\item{p_visibility}{list of probabilities of visibility (show-up)}

\item{p_service}{probability of service utilization in hidden population}

\item{directed}{logical, whether links are directed or undirected}
}
\value{

}
\description{
Simulate signle population with given network structure
}
\examples{
\dontrun{
population_network(
  # total population size for one study
  N = 1000,
  # number of groups
  # (K-th group is hidden population we are interested in)
  K = 2,
  # probability of membership in each of the groups (prev_K[K] is the true prevalence)
  prev_K = c(known = .3, hidden = .1),
  # correlation matrix of group memberships
  rho_K = .05,
  # block edge probabilities depending on group memberships
  # 1 - list of in- and out-group probability of links for each group
  # 2 - probability of link between in- and out-group members
  p_edge_within = list(known = c(0.05, 0.05), hidden = c(0.05, 0.9)),
  p_edge_between = list(known = 0.05, hidden = 0.01),
  # probability of visibility (show-up) for each group
  p_visibility = list(hidden = 1, known = 1),
  # probability of service utilization in hidden population
  # for service multiplier
  p_service = 0.3)
}

}
