% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet_cpp}
\alias{rdirichlet_cpp}
\title{helper to generate single draw from dirichlet distribution}
\usage{
rdirichlet_cpp(alpha)
}
\arguments{
\item{alpha}{double vector of alpha parameters}
}
\value{
a double vector containing a single draw from a dirichlet distribution
}
\description{
helper to generate single draw from dirichlet distribution
}
\keyword{internal}
