% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{read_estimators}
\alias{read_estimators}
\title{Read Study Estimators from a Spreadsheet generated by \code{read_gs()}}
\usage{
read_estimators(study_df, append_list)
}
\arguments{
\item{study_df}{Data frame that contains each study information. Usually generated by \code{hiddenmeta::read_gs()}.}

\item{append_list}{List with study designs to append that already contains at the minimum population and sampling information from the \code{study_df}.}
}
\description{
Read Study Estimators from a Spreadsheet generated by \code{read_gs()}
}
