% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{read_gs}
\alias{read_gs}
\title{Create a Spreadsheet with Study Designs from Google Spreadsheet}
\usage{
read_gs(
  ss = "1HwMM6JwoGALLMTpC8pQRVzaQdD2X71jpZNhfpKTnF8Y",
  sheet = "study_params_readable"
)
}
\arguments{
\item{ss}{Character string giving ID of the Google spreadsheet}

\item{sheet}{Character string giving name of the sheet in the Google spreadsheet to read}
}
\description{
Create a Spreadsheet with Study Designs from Google Spreadsheet
}
