% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{read_populations}
\alias{read_populations}
\title{Read Study Population Designs from a Spreadsheet generated by \code{read_gs()}}
\usage{
read_populations(study_df, append_list = NULL)
}
\arguments{
\item{study_df}{Data frame that contains each study information. Usually generated by \code{hiddenmeta::read_gs()}.}

\item{append_list}{List with study designs to append. Defaults to \code{NULL}, which will generate new study list from the \code{study_df}.}
}
\description{
Read Study Population Designs from a Spreadsheet generated by \code{read_gs()}
}
