% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{read_single_study_params}
\alias{read_single_study_params}
\title{Get Parameters and Design Features of Specific Study}
\usage{
read_single_study_params(study = "johnjay_tanzania", ss, sheet)
}
\arguments{
\item{study}{Character string giving name of the study to pull. Have to match the names provided in Google Spreadsheet}

\item{ss}{Character string giving ID of the Google spreadsheet}

\item{sheet}{Character string giving name of the sheet in the Google spreadsheet to read}
}
\value{

}
\description{
Get Parameters and Design Features of Specific Study
}
\examples{

}
