% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{read_study_params}
\alias{read_study_params}
\title{Read Study Designs from Google Spreadsheet}
\usage{
read_study_params(
  ss = "1HwMM6JwoGALLMTpC8pQRVzaQdD2X71jpZNhfpKTnF8Y",
  sheet = "study_params_readable"
)
}
\arguments{
\item{ss}{Character string giving ID of the Google spreadsheet}

\item{sheet}{Character string giving name of the sheet in the Google spreadsheet to read}
}
\value{

}
\description{
Read Study Designs from Google Spreadsheet
}
\examples{

}
