% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rep_each}
\alias{rep_each}
\title{cpp implementation of R rep with each argument}
\usage{
rep_each(x, n)
}
\arguments{
\item{x}{integer vector of elements to be repeated}

\item{n}{integer number of repetitions}
}
\value{
integer vector with each element in x repeated n times in order
}
\description{
cpp implementation of R rep with each argument
}
\keyword{internal}
