% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rep_times}
\alias{rep_times}
\title{cpp implementation of standard R rep}
\usage{
rep_times(x, n)
}
\arguments{
\item{x}{integer vector to be repeated}

\item{n}{integer number of repetitions}
}
\value{
integer vector x repeated n times
}
\description{
cpp implementation of standard R rep
}
\keyword{internal}
