% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_pps.R
\name{sample_pps}
\alias{sample_pps}
\title{Draw probability sample proportional to size (PPS) from single study}
\usage{
sample_pps(
  data,
  sample_label = "pps",
  drop_nonsampled = FALSE,
  target_n_pps = 400,
  n_clusters = target_n_pps,
  sampling_frame = NULL,
  strata = NULL,
  cluster = NULL,
  weights_type = c("absolute", "relative")
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sample_label}{character string that is used as prefix for all variables generated by proportional sampling. Default is 'pps'}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{target_n_pps}{target size of proportional sample}

\item{n_clusters}{number of clusters}

\item{sampling_frame}{character vector containing all binary vectors identifying sampling frame}

\item{strata}{character vector containing name(s) of stratifying variables. Currently not implemented}

\item{cluster}{character vector containing name(s) of all cluster variables}

\item{weights_type}{character string giving the type of weights to compute. Can be one of "absolute" or "relative". Currently only absolute weights are calculated}
}
\value{
Population or sample data frame for single study with PPS sample characteristics added
 \describe{
  \item{[sample_label]}{Sampling indicator}
  \item{[sample_label]_frame}{Indicator for sampling frame (respondents with 0 cannot be enrolled)}
  \item{[sample_label]_strata}{ID of respondent's strata}
  \item{[sample_label]_strata_prop}{Proportion of sampling frame in the strata to which respondent belongs}
  \item{[sample_label]_cluster}{ID of respondent's cluster}
  \item{[sample_label]_cluster_prop}{Proportion of strata in the cluster to which respondent belongs}
  \item{[sample_label]_weight}{Sampling weights}
 }
}
\description{
Sampling handler for drawing proportional sample with given characteristics from individual study population
}
