% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rds.R
\name{sample_rds}
\alias{sample_rds}
\title{Title}
\usage{
sample_rds(
  data,
  assignment_variable = "RDS",
  n_seed_hidden,
  n_seed_other,
  n_coupons,
  target_size_rds
)
}
\arguments{
\item{target_size_rds}{}
}
\value{

}
\description{
Title
}
