% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_tls.R
\name{sample_tls_tidy}
\alias{sample_tls_tidy}
\title{Draw time-location (TLS) sample from single study}
\usage{
sample_tls_tidy(
  data,
  sampling_variable = "tls",
  drop_nonsampled = FALSE,
  hidden_var = "hidden",
  target_n_clusters,
  target_cluster_type = c("prop", "fixed"),
  target_per_cluster,
  clusters
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sampling_variable}{character string that is used as prefix for all variables generated by TLS sampling. Default is 'tls'}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{hidden_var}{character string specifying hidden variable name (associated probability of visibility should be named \code{p_visible_[hidden_var]}). Defaults to "hidden" for the simulations}

\item{target_n_clusters}{target number of clusters (time-locations). Clusters are always sampled proportionally to their size in terms of number of hidden population members}

\item{target_cluster_type}{character string specifying the type of TLS sampling within each location. Either "prop" in which case \code{target_per_cluster} should give a share or "fixed" in which case \code{target_per_cluster} should be integer. Default is proportional sampling within clusters}

\item{target_per_cluster}{numeric target for within cluster. Either share for proportional sampling or integer for fixed sampling. If in any cluster fixed number of units required for sampling is larger than the number of units in cluster, the whole cluster is sampled and the warning is produced}

\item{clusters}{character string containing names of all locality names in the study population data frame}
}
\value{
Population or sample data frame for single study with TLS sample characteristics added
 \describe{
  \item{[sampling_variable]}{Sampling indicator}
  \item{[sampling_variable]_cluster}{Time-locations at which subject was encountered first}
  \item{[sampling_variable]_loc_present}{Sampled time-locations at which subject is present}
  \item{[sampling_variable]_loc_sampled}{Sampled time-locations at which subject was sampled}
  \item{[sampling_variable]_locs_sampled}{Sampled time-locations}
  \item{[sampling_variable]_weight}{Sampling weight without visibility}
  \item{[sampling_variable]_weight_visible}{Sampling weight with visibility}
 }
}
\description{
Sampling handler for drawing TLS sample with given characteristics from individual study population
}
