% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_network.R
\name{sim_ergm_network}
\alias{sim_ergm_network}
\title{Handler for Simulation of Network Using ERGM}
\usage{
sim_ergm_network(fit, type_function)
}
\arguments{
\item{fit}{object of ergm class produced by fitting model using \code{ergm} function}

\item{type_function}{function of igraph object \code{g} that takes existing vertex attributes and transforms them into \code{type} attribute in the binary coded format (consists of 0's and 1's only)}
}
\value{
igraph network object with vertex attribute \code{type} in the binary coded format (consists of 0's and 1's only)
}
\description{
Draw a Simulation of Network Using Distribution Of An Exponential Family Random Graph Model Based on Observed Data
}
\examples{
\dontrun{
library(network)
library(ergm)

data('faux.magnolia.high')

fit <- ergm(faux.magnolia.high ~
              edges +
              gwesp(0.25,fixed=T) +
              nodematch("Grade") + nodematch("Sex"),
            control = control.ergm(MCMC.interval = 10000),
            verbose = F)

sim_ergm_network(fit = fit,
                 type_function = function(g) {
                   paste0( as.integer(igraph::vertex_attr(g)$Sex == "M"),
                           as.integer(igraph::vertex_attr(g)$Grade == 9))})

}

}
