% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{table_cpp}
\alias{table_cpp}
\title{cpp implementation of R table}
\usage{
table_cpp(x)
}
\arguments{
\item{x}{an integer vector of elements to be counted}
}
\value{
an integer vector of counts of unique elements in x (sorted in ascending order of elements in x)
}
\description{
cpp implementation of R table
}
\keyword{internal}
