% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rds.R
\name{get_hpop_rds}
\alias{get_hpop_rds}
\title{Draw RDS samples from multiple study populations}
\usage{
get_hpop_rds(data, rds_args)
}
\arguments{
\item{data}{pass-through populations data frame}

\item{rds_args}{named list of named lists of arguments to \code{sample_rds()} (the set of arguments can be excessive)}
}
\value{

}
\description{
Sampling handler for drawing RDS sample with given characteristics from multiple study populations. The RDS sampling characteristics can vary across studies
}
