% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_estimands.R
\name{get_meta_estimands}
\alias{get_meta_estimands}
\title{Get meta-analysis estimands for specific study-level estimand}
\usage{
get_meta_estimands(
  data,
  study_estimand = "hidden_size",
  samp_est_benchmark = "pps_ht"
)
}
\arguments{
\item{data}{pass-through meta population}

\item{study_estimand}{study-level estimand for meta-analysis estimands}

\item{samp_est_benchmark}{within-study benchmark sampling-estimator strategy}
}
\value{

}
\description{
Get meta-analysis estimands for specific study-level estimand
}
