% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_estimator.R
\name{get_meta_estimates}
\alias{get_meta_estimates}
\title{Use Stan model to estimate bias and estimands}
\usage{
get_meta_estimates(
  data,
  sampling_variable = "meta",
  which_estimand = "hidden_size",
  benchmark = list(sample = "pps", estimator = "ht"),
  parallel = FALSE
)
}
\arguments{
\item{data}{pass-through meta population or meta sample data frame}

\item{sampling_variable}{name of variable storing meta analysis sampling information}

\item{which_estimand}{name of study level estimand for meta analysis}

\item{benchmark}{named list of length 2 giving benchmark sampling-estimator pair (only accepts one value across studies for now)}

\item{parallel}{logical. Whether to parallelize design simulations}
}
\value{

}
\description{
Use Stan model to estimate bias and estimands
}
