% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_estimator.R
\name{get_meta_estimates}
\alias{get_meta_estimates}
\title{Use Stan model to estimate bias and estimands}
\usage{
get_meta_estimates(
  data,
  sample_label = "meta",
  benchmark = list(sample = "pps", estimator = "ht"),
  hidden_prior = list(mean = 0.3, sd = 0.2),
  bias_prior = list(mean = 1, sd = 3),
  stan_handler = get_meta_stan4,
  control_params = list(iter = 8000, chains = 8, thin = 10, seed = 872312, cores = 1),
  return_stanfit = FALSE
)
}
\arguments{
\item{data}{pass-through meta population or meta sample data frame}

\item{sample_label}{name of variable storing meta analysis sampling information}

\item{benchmark}{named list of length 2 giving benchmark sampling-estimator pair (only accepts one value across studies for now)}

\item{hidden_prior}{list of two hyperpriors, on means and standard errors of each included sampling-estimator pairs. Names of list objects should be "mean" and "se". If one number provided for a hyperprior it gets expanded to all sampling-estimator pairs}

\item{stan_handler}{function that takes stan_data as input and produces compilable stan model object}

\item{control_params}{list of additional parameters to pass to Stan fit function. These can include number of iterations, chains, thinning, seed and number of cores to use}

\item{return_stanfit}{logical value identifying whether the Stan model fit object should be returned in addition to estimates data frame}

\item{rel_bias_prior}{list of two hyperpriors, on means and standard errors of relative bias. Names of list objects should be "mean" and "se".}
}
\value{
Data frame of meta level estimates and pertaining estimand names
}
\description{
Use Stan model to estimate bias and estimands
}
