% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_population.R
\name{get_meta_population}
\alias{get_meta_population}
\title{Simulate meta study population using simulation of several studies}
\usage{
get_meta_population(multi_design, n_sim = 2, parallel = FALSE)
}
\arguments{
\item{multi_design}{design class object declaring full design of the set of studies with all possible sampling-estimator pairs}

\item{n_sim}{number of simulations to use}

\item{parallel}{logical. Whether to parallelize design simulations}
}
\value{

}
\description{
Simulate meta study population using simulation of several studies
}
