% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_meta.R
\name{get_meta_stan}
\alias{get_meta_stan}
\title{Parse stan model for meta analysis}
\usage{
get_meta_stan(stan_data)
}
\arguments{
\item{stan_data}{list of Stan model data inputs}
}
\value{

}
\description{
Parse stan model for meta analysis
}
