% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_meta.R
\name{get_multi_estimands}
\alias{get_multi_estimands}
\title{Calculate estimands from multiple study populations}
\usage{
get_multi_estimands(data, estimands_args)
}
\arguments{
\item{data}{pass-through populations data frame}

\item{estimands_args}{named list of named lists of arguments to estimand function}
}
\value{
Data frame with estimands and estimand labels
}
\description{
Handler for calculating estimand(s) from multiple study populations
}
