% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_multi.R
\name{get_multi_samples}
\alias{get_multi_samples}
\title{Draw samples from multiple study populations according to proposed strategies}
\usage{
get_multi_samples(data, study_designs)
}
\arguments{
\item{data}{pass-through populations data frame}

\item{study_designs}{named list of arguments to sampling strategies used in each study.}
}
\value{
Meta tibble data frame with columns
\itemize{
 \item{"study"}{Character. Study identifier}
 \item{"sample"}{Character. Sampling strategy identifier (possible multiple per study}
 \item{"population"}{List of tibble data frames. For each combination of study and sampling strategy the data frame includes full population information and respective sampling strategy information}
}
}
\description{
Handler for drawing sample(s) with given characteristics from multiple study populations. The sampling characteristics can vary across and within studies
}
