% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_populations}
\alias{get_populations}
\title{Simulate population with given network structure for multiple studies with varying designs}
\usage{
get_populations(pop_args)
}
\arguments{
\item{pop_args}{named list of named lists of arguments to population_network()}
}
\value{

}
\description{
Simulate population with given network structure for multiple studies with varying designs
}
