% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_import.R
\name{get_required_data}
\alias{get_required_data}
\title{Generate Data Frame With Required Data by Study}
\usage{
get_required_data(
  multi_study_data,
  study_map = c(`Brazil (FF)` = "ff_brazil", `Tunisia (UMass)` = "umass_tunisia",
    `Brazil (Stanford)` = "stanford_brazil", `Pakistan (JHU)` = "jhu_pakistan", `Costa
    Rica (John Jay)` = "johnjay_costarica", `Tanzania (John Jay)` = "johnjay_tanzania",
    `Morocco (NORC)` = "norc_marocco", `USA (RTI)` = "rti_usa")
)
}
\arguments{
\item{multi_study_data}{tibble with two columns, study (character identifier of the study) and population (list of population data frames/tibbles). This can be directly generated by multi-study design declaration}

\item{study_map}{named character vector of mapping between study labels and names used in the plot}
}
\value{

}
\description{
Generate Data Frame With Required Data by Study
}
\examples{

}
