% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_chords}
\alias{get_study_est_chords}
\title{Chords estimatior}
\usage{
get_study_est_chords(data)
}
\arguments{
\item{data}{pass-through population data frame}
}
\value{
Data frame of HT estimates for single study
}
\description{
Chords estimatior
}
\references{
Berchenko, Yakir, Jonathan D. Rosenblatt, and Simon D. W. Frost. “Modeling and Analyzing Respondent-Driven Sampling as a Counting Process.” Biometrics 73, no. 4 (2017): 1189–98. \url{https://doi.org/10.1111/biom.12678}.
}
