% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_chords}
\alias{get_study_est_chords}
\title{Chords population size estimatior by Berchenko, Rosenblatt and Frost}
\usage{
get_study_est_chords(
  data,
  type = c("mle", "integrated", "jeffreys"),
  seed_condition = "rds_from == -999",
  prefix = "rds",
  n_boot = 100,
  parallel_boot = FALSE,
  label = "chords"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{type}{a character vector with the type of estimation. Can be one of \code{mle}, \code{integrated}, or \code{jeffreys}. See \code{?chords::Estimate.b.k} and the original paper from the references for details}

\item{seed_condition}{character string containing condition to define seeds. Defaults to "rds_from == -999" that applies to simulated RDS samples}

\item{prefix}{character string prefix used for RDS sample variables}

\item{n_boot}{number of bootstrap resamples}

\item{parallel_boot}{logical, whether to compute bootstrap samples in parallel using \code{foreach} package}

\item{label}{character string describing the estimator}
}
\value{
Data frame of Chords estimates for a single study with RDS sample
}
\description{
Chords population size estimatior by Berchenko, Rosenblatt and Frost
}
\references{
{Berchenko, Yakir, Jonathan D. Rosenblatt, and Simon D. W. Frost. “Modeling and Analyzing Respondent-Driven Sampling as a Counting Process.” Biometrics 73, no. 4 (2017): 1189–98. \url{https://doi.org/10.1111/biom.12678}.}

{Salganik, Matthew J. "Variance estimation, design effects, and sample size calculations for respondent-driven sampling." Journal of Urban Health 83, no. 1 (2006): 98. \url{https://doi.org/10.1007/s11524-006-9106-x}.}
}
