% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_gnsum}
\alias{get_study_est_gnsum}
\title{Generalized NSUM estimatior}
\usage{
get_study_est_gnsum(data)
}
\arguments{
\item{data}{pass-through population data frame}
}
\value{
Data frame of HT estimates for single study
}
\description{
Generalized NSUM estimatior
}
