% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_ht}
\alias{get_study_est_ht}
\title{Horvitz-Thompson prevalence estimator using weighted regression}
\usage{
get_study_est_ht(data, prefix = "pps", label = "ht")
}
\arguments{
\item{data}{pass-through population data frame}

\item{prefix}{character prefix used for sampling variables (has to include \code{[prefix]_weights})}

\item{label}{character string describing the estimator}
}
\value{
Data frame of HT estimates for a single study
}
\description{
Horvitz-Thompson prevalence estimator using weighted regression
}
