% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_ht}
\alias{get_study_est_ht}
\title{Horvitz-Thompson prevalence estimator using weighted regression}
\usage{
get_study_est_ht(data, pps_prefix = "pps")
}
\arguments{
\item{data}{pass-through population data frame}

\item{pps_prefix}{character prefix used for RDS sample variables}
}
\value{
Data frame of HT estimates for a single study
}
\description{
Horvitz-Thompson prevalence estimator using weighted regression
}
