% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_nsum}
\alias{get_study_est_nsum}
\title{NSUM estimatior}
\usage{
get_study_est_nsum(data)
}
\arguments{
\item{data}{pass-through population data frame}
}
\value{
Data frame of HT estimates for single study
}
\description{
NSUM estimatior
}
