% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_nsum}
\alias{get_study_est_nsum}
\title{NSUM estimatior}
\usage{
get_study_est_nsum(
  data,
  pps_prefix = "pps",
  known_pattern = "known",
  hidden_pattern = "hidden_visible",
  degree_ratio = 1,
  transmission_rate = 1
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{pps_prefix}{character prefix used for PPS sample variables}

\item{known_pattern}{character prefix for known population variables}

\item{hidden_pattern}{character prefix for hidden population variable}

\item{degree_ratio}{numeric value between 0 and 1 representing degree ratio}

\item{transmission_rate}{numeric value between 0 and 1 representing information transmission rate}
}
\value{
Data frame of NSUM estimates for a single study with PPS sample
}
\description{
NSUM estimatior
}
\references{
Dennis M. Feehan, Matthew J. Salganik. “The networkreporting package.” (2014). \url{https://cran.r-project.org/package=networkreporting}.
}
