% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_nsum}
\alias{get_study_est_nsum}
\title{NSUM estimatior}
\usage{
get_study_est_nsum(
  data,
  known,
  hidden,
  survey_design = ~pps_cluster + strata(pps_stata),
  degree_ratio = 1,
  transmission_rate = 1,
  n_boot = 1000,
  prefix = "pps",
  label = "nsum"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{known}{character vector containing names of known groups}

\item{hidden}{character vector containing names of hidden groups}

\item{survey_design}{a formula describing the design of the survey}

\item{degree_ratio}{numeric value between 0 and 1 representing degree ratio}

\item{transmission_rate}{numeric value between 0 and 1 representing information transmission rate}

\item{n_boot}{number of bootstrap resamples}

\item{prefix}{character prefix used for PPS sample variables}

\item{label}{character string describing the estimator}
}
\value{
Data frame of NSUM estimates for a single study with PPS sample
}
\description{
NSUM estimatior
}
\references{
Dennis M. Feehan, Matthew J. Salganik. “The networkreporting package.” (2014). \url{https://cran.r-project.org/package=networkreporting}.
Dennis M. Feehan, Matthew J. Salganik. “The surveybootstrap package.” (2016). \url{https://cran.r-project.org/package=surveybootstrap}.
}
