% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_nsum}
\alias{get_study_est_nsum}
\title{NSUM estimatior}
\usage{
get_study_est_nsum(
  data,
  prefix = "pps",
  known = c("known", "known_2", "known_3"),
  hidden = "hidden_visible_out",
  degree_ratio = 1,
  transmission_rate = 1,
  boot_n = 1000,
  label = "nsum"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{prefix}{character prefix used for PPS sample variables}

\item{known}{character vector containing names of known groups}

\item{hidden}{character vector containing names of hidden groups}

\item{degree_ratio}{numeric value between 0 and 1 representing degree ratio}

\item{transmission_rate}{numeric value between 0 and 1 representing information transmission rate}

\item{label}{character string describing the estimator}
}
\value{
Data frame of NSUM estimates for a single study with PPS sample
}
\description{
NSUM estimatior
}
\references{
Dennis M. Feehan, Matthew J. Salganik. “The networkreporting package.” (2014). \url{https://cran.r-project.org/package=networkreporting}.
}
