% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_recapture}
\alias{get_study_est_recapture}
\title{Mark-recapture estimator for closed population}
\usage{
get_study_est_recapture(
  data,
  capture_vars = NULL,
  capture_parse = NULL,
  sample_condition = NULL,
  model = "Mt",
  hidden_variable = "hidden",
  label = "recapture"
)
}
\arguments{
\item{data}{pass-through population data frame that contains capture indicators}

\item{capture_vars}{character vector giving names of variables with capture indicators}

\item{capture_parse}{character string giving expression to evaluation of which produces character vector giving names of variable with capture indicators. Defaults to \code{NULL}. This is useful when capture variables are stored in one column (e.g. if using TLS sampled locations for recapture indicators)}

\item{sample_condition}{character string with condition if the capture-recapture conducted on subsample of population (e.g. tls sample only)}

\item{model}{character string giving capture-recapture Log-Linear model to estimate}

\item{hidden_variable}{character string giving indicator of hidden population membership}

\item{label}{character string describing the estimator}
}
\value{
Data frame of HT estimates for single study
}
\description{
Mark-recapture estimator for closed population
}
\references{
Louis-Paul Rivest, Sophie Baillargeon. “The Rcapture package.” (2019). \url{https://cran.r-project.org/package=Rcapture}.
}
