% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_service}
\alias{get_study_est_service}
\title{Service multiplier estimator}
\usage{
get_study_est_service(data, label = "service")
}
\arguments{
\item{data}{pass-through population data frame}

\item{label}{character string describing the estimator}
}
\value{
Data frame of HT estimates for single study
}
\description{
Service multiplier estimator
}
