% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_sspse}
\alias{get_study_est_sspse}
\title{SS-PSE population size estimator by Handcock, Gile and Mar}
\usage{
get_study_est_sspse(
  data,
  rds_prefix = "rds",
  prior_mean = 150,
  prior_sd = 50,
  n_coupons = 3,
  mcmc_params = list(interval = 1, burnin = 1000, samplesize = 100),
  label = "sspse"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{rds_prefix}{character prefix used for RDS sample variables}

\item{prior_mean}{the mean of the prior distribution on the population size for SS-PSE estimation}

\item{prior_sd}{the standard deviation of the prior distribution on the population size for SS-PSE estimation}

\item{n_coupons}{The number of recruitment coupons distributed to each enrolled subject (i.e. the maximum number of recruitees for any subject). By default it is taken by the attribute or data, else the maximum recorded number of coupons.}

\item{mcmc_params}{parameters for MCMC sampling}

\item{label}{character string describing the estimator}
}
\value{
Data frame of SS-PSE estimates for a single study
}
\description{
SS-PSE population size estimator by Handcock, Gile and Mar
}
\references{
Handcock, Mark S., Krista J. Gile, and Corinne M. Mar. “Estimating Hidden Population Size Using Respondent-Driven Sampling Data.” Electronic Journal of Statistics 8, no. 1 (2014): 1491–1521. \url{https://doi.org/10.1214/14-EJS923}.
}
