% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_sspse}
\alias{get_study_est_sspse}
\title{SS-PSE population size estimator by Handcock, Gile and Mar}
\usage{
get_study_est_sspse(
  data,
  prior_mean = 0.1 * nrow(data),
  n_coupons = 3,
  total = 2000,
  mcmc_params = list(interval = 5, burnin = 2000, samplesize = 500),
  additional_params = list(),
  prefix = "rds",
  label = "sspse"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{prior_mean}{the mean of the prior distribution on the population size for SS-PSE estimation}

\item{n_coupons}{The number of recruitment coupons distributed to each enrolled subject (i.e. the maximum number of recruitees for any subject). By default it is taken by the attribute or data, else the maximum recorded number of coupons.}

\item{total}{integer giving the total size of population}

\item{mcmc_params}{named list of parameters passed to \code{sspse::posteriorsize} for MCMC sampling,}

\item{additional_params}{named list of additional parameter passed to \code{sspse::posteriorsize} . If empty \code{sspse::posteriorsize} uses default parameters.}

\item{prefix}{character prefix used for RDS sample variables}

\item{label}{character string describing the estimator}
}
\value{
Data frame of SS-PSE estimates for a single study
}
\description{
SS-PSE population size estimator by Handcock, Gile and Mar
}
\references{
{Handcock, Mark S., Krista J. Gile, and Corinne M. Mar. “Estimating Hidden Population Size Using Respondent-Driven Sampling Data.” Electronic Journal of Statistics 8, no. 1 (2014): 1491–1521. \url{https://doi.org/10.1214/14-EJS923}.}
}
