% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimator_sspse.R
\name{get_study_est_sspse}
\alias{get_study_est_sspse}
\title{Title}
\usage{
get_study_est_sspse(data, prior_median = 150)
}
\arguments{
\item{data}{pass-through population data frame}

\item{prior_median}{prior median of hidden population size for SS-PSE estimation}
}
\value{

}
\description{
Title
}
