% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimators.R
\name{get_study_est_sspse}
\alias{get_study_est_sspse}
\title{SS-PSE population size estimator by Handcock, Gile and Mar}
\usage{
get_study_est_sspse(data, prior_median = 150, rds_prefix = "rds")
}
\arguments{
\item{data}{pass-through population data frame}

\item{prior_median}{prior median of hidden population size for SS-PSE estimation}

\item{rds_prefix}{character prefix used for RDS sample variables}
}
\value{
Data frame of SS-PSE estimates for a single study
}
\description{
SS-PSE population size estimator by Handcock, Gile and Mar
}
\references{
Handcock, Mark S., Krista J. Gile, and Corinne M. Mar. “Estimating Hidden Population Size Using Respondent-Driven Sampling Data.” Electronic Journal of Statistics 8, no. 1 (2014): 1491–1521. \url{https://doi.org/10.1214/14-EJS923}.
}
