% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_estimands.R
\name{get_study_estimands}
\alias{get_study_estimands}
\title{Get individual study estimands}
\usage{
get_study_estimands(data)
}
\arguments{
\item{data}{pass-through population data frame}
}
\value{

}
\description{
Get individual study estimands
}
