% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_population.R
\name{get_study_population}
\alias{get_study_population}
\title{Simulate single population with given network structure}
\usage{
get_study_population(
  N = 1000,
  K = 2,
  prev_K = c(known = 0.3, hidden = 0.1),
  rho_K = 0.05,
  p_edge_within = list(known = c(0.05, 0.05), hidden = c(0.05, 0.9)),
  p_edge_between = list(known = 0.05, hidden = 0.01),
  p_visibility = list(hidden = 0.7, known = 0.99),
  add_groups = list(p_service = 0.3, loc_A = 0.01, loc_B = 0.01),
  directed = FALSE
)
}
\arguments{
\item{N}{number of units in population}

\item{K}{number of groups}

\item{prev_K}{named numeric vector of prevalence for each group with last group being hidden population}

\item{rho_K}{numeric vector of correlations in group memberships}

\item{p_edge_within}{named list of numeric vectors giving probability of link between in-group members and out-group members for each of groups. The order of objects in list have to follow the order of \code{prev_K}}

\item{p_edge_between}{named list of numeric values giving probability of link between in- and out-group member for each of groups. The order of objects in list have to follow the order of \code{prev_K}}

\item{p_visibility}{named list of visibility tendencies by group. This is used as mean of Beta distribution (with SD = 0.09) to generate probability of being recognized as member of group, being sampled as seed, etc. The order of objects in list have to follow the order of \code{prev_K}}

\item{add_groups}{named list of probabilities of additional group memberships. Examples include probability of service utilization (for service multiplier), going to particular location (for TLS), etc.}

\item{directed}{logical, whether links are directed or undirected}
}
\value{

}
\description{
Simulate single population with given network structure
}
\examples{
\dontrun{
pop_network(
  # total population size for one study
  N = 1000,
  # number of groups
  # (K-th group is hidden population we are interested in)
  K = 2,
  # probability of membership in each of the groups (prev_K[K] is the true prevalence)
  prev_K = c(known = .3, hidden = .1),
  # correlation matrix of group memberships
  rho_K = .05,
  # block edge probabilities depending on group memberships
  # 1 - list of in- and out-group probability of links for each group
  # 2 - probability of link between in- and out-group members
  p_edge_within = list(known = c(0.05, 0.05), hidden = c(0.05, 0.9)),
  p_edge_between = list(known = 0.05, hidden = 0.01),
  # probability of visibility (show-up) for each group
  p_visibility = list(hidden = 1, known = 1),
  # probability of service utilization in hidden population
  # for service multiplier
  p_service = 0.3)
}

}
