% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_linktrace.R
\name{lt_permute}
\alias{lt_permute}
\title{Link-tracing sample permutation}
\usage{
lt_permute(data)
}
\arguments{
\item{data}{pass through sample data frame}
}
\value{
list of IDs for observations in each sampling wave
}
\description{
Link-tracing sample permutation
}
