% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_pps.R
\name{sample_pps}
\alias{sample_pps}
\title{Draw proportional sample (PPS) from single study}
\usage{
sample_pps(
  data,
  sampling_variable = "prop",
  drop_nonsampled = FALSE,
  target_n_pps = 400,
  group_pattern = "^known|^hidden$"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sampling_variable}{character string that is used as prefix for all variables generated by proportional sampling. Default is 'prop'}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{target_n_pps}{target size of proportional sample}

\item{group_pattern}{character string containing regular expression to match all groups used for proportional sampling}
}
\value{
Population or sample data frame for single study with PPS sample characteristics added
}
\description{
Sampling handler for drawing proportional sample with given characteristics from individual study population
}
