% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_prop.R
\name{sample_prop}
\alias{sample_prop}
\title{Draw proportional sample from single study}
\usage{
sample_prop(
  data,
  sampling_variable = "prop",
  drop_nonsampled = FALSE,
  target_size_prop = 400
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sampling_variable}{character string that is used as prefix for all variables generated by RDS sampling (sample identifier, recruiter ID, wave, time of show-up)}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{target_size_prop}{target size of proportional sample}
}
\value{

}
\description{
Sampling handler for drawing proportional sample with given characteristics from individual study population
}
