% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_lts.R
\name{sample_rds}
\alias{sample_rds}
\title{Draw respondent-driven sample (RDS) or link-tracing (LTS) sample from single study using data.table}
\usage{
sample_rds(
  data,
  sample_label = "lts",
  hidden_var,
  sampling_frame = hidden_var,
  n_coupons,
  target_type = c("sample", "waves"),
  target_n_rds,
  n_seed,
  n_waves = NULL,
  add_seeds = NULL,
  arrival_rate = 5,
  drop_nonsampled = FALSE
)
}
\arguments{
\item{data}{pass-through population data frame.}

\item{sample_label}{character string that is used as prefix for all variables generated by RDS sampling (sample identifier, recruiter ID, wave, time of show-up).}

\item{hidden_var}{character string specifying hidden variable name (associated probability of visibility should be named \code{p_visible_[hidden_var]}).}

\item{sampling_frame}{character string indicating either \code{"all"} or name of variable to use for sampling if non-hidden group members are allowed. Default is the same as \code{hidden_var} for hidden group only sampling.}

\item{n_coupons}{number of unique coupons given to each study participant. Defaults to \code{3}. If \code{n_coupons == Inf}, all links are traced and total number of available coupons is assumed to be equal to maximum degree in population.}

\item{target_type}{one of \code{"sample"} or \code{"waves"}. If \code{"sample"} RDS sampling proceeds until either sample size of \code{target_n_rds} is reached or all available links are depleted. If \code{"waves"} in addition number of waves is restricted to be \code{n_waves}.}

\item{target_n_rds}{integer vector giving target sample size(s). If \code{length(target_n_rds) > 1}, it is assumed that more than one RDS sample is requested for link-tracing.}

\item{n_seed}{integer vector giving number(s) of seeds to start sampling from. If \code{length(n_seed) > 1}, it is assumed that more than one RDS sample is requested for link-tracing. Has to match length of \code{target_n_rds}.}

\item{n_waves}{integer vector giving number(s) of waves allowed. Defaults to \code{NULL}}

\item{add_seeds}{numeric indicating how many seeds to add at a time if target sample size is not reached with initial seeds. Additional seeds are randomly drawn from non-sampled hidden population members. Defaults to \code{NULL} that does not allow adding seeds.}

\item{arrival_rate}{numeric rate of respondent arrival per interval of time (e.g. per hour or day). Defaults to \code{5}.}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Defaults to \code{FALSE}.}
}
\value{
Population or sample data frame for single study with RDS sample characteristics added
 \describe{
  \item{[sample_label]}{Sampling indicator}
  \item{[sample_label]_from}{ID of respondent who enrolled current respondent}
  \item{[sample_label]_t}{Time at which respondent was enrolled}
  \item{[sample_label]_wave}{Number of steps respondent is away from the seed (seeds are wave 0)}
  \item{[sample_label]_own_coupon}{ID of the coupon with which respondent was enrolled}
  \item{[sample_label]_coupon_[1-9]}{IDs of the coupons that were given to respondent for enrollment from their network}
 }
}
\description{
Sampling handler for drawing RDS sample with given characteristics from individual study population
}
