% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_rds.R
\name{sample_rds}
\alias{sample_rds}
\title{Draw respondent-driven sample (RDS) sample from single study}
\usage{
sample_rds(
  data,
  sampling_variable = "rds",
  hidden_var,
  target_type = c("sample", "waves"),
  n_seed,
  n_coupons,
  target_n_rds,
  add_seeds = NULL,
  arrival_rate = 0.5,
  drop_nonsampled = FALSE
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sampling_variable}{character string that is used as prefix for all variables generated by RDS sampling (sample identifier, recruiter ID, wave, time of show-up)}

\item{hidden_var}{character string specifying hidden variable name (associated probability of visibility should be named \code{p_visible_[hidden_var]}). Defaults to "hidden" for the simulations}

\item{target_type}{one of 'sample' or 'waves'}

\item{n_seed}{number of seeds randomly drawn from members of hidden population (group K)}

\item{n_coupons}{number of unique coupons given to each study participant}

\item{target_n_rds}{numeric target size of RDS sample. If \code{target_type = "sample"}, this gives maximum number of respondents to be sampled (right now the RDS network can also end before reaching sample size target). If \code{target_type = "waves"}, this gives maximum number of waves of recruitment allowed}

\item{add_seeds}{numeric indicating how many seeds to add at a time if target sample size is not reached with initial seeds. Additional seeds are randomly drawn from non-sampled hidden population members. Defaults to \code{NULL} that does not allow adding seeds}

\item{arrival_rate}{numeric rate of respondent arrival per unit of time (e.g. per day). Defaults to .5}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}
}
\value{
Population or sample data frame for single study with RDS sample characteristics added
 \describe{
  \item{[sampling_variable]}{Sampling indicator}
  \item{[sampling_variable]_from}{ID of respondent who enrolled current respondent}
  \item{[sampling_variable]_t}{Time at which respondent was enrolled}
  \item{[sampling_variable]_wave}{Number of steps respondent is away from the seed (seeds are wave 0)}
  \item{[sampling_variable]_own_coupon}{ID of the coupon with which respondent was enrolled}
  \item{[sampling_variable]_coupon_[1-9]}{IDs of the coupons that were given to respondent for enrollment from their network}
 }
}
\description{
Sampling handler for drawing RDS sample with given characteristics from individual study population
}
\examples{
\dontrun{
sample_rds(data = get_pop_network())
}

}
