% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_rds.R
\name{sample_rds}
\alias{sample_rds}
\title{Draw RDS sample from single study}
\usage{
sample_rds(
  data,
  sampling_variable = "rds",
  drop_nonsampled = FALSE,
  n_seed = 10,
  n_coupons = 3,
  target_type = c("sample", "waves"),
  target_n = 40
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sampling_variable}{character string that is used as prefix for all variables generated by RDS sampling (sample identifier, recruiter ID, wave, time of show-up)}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{n_seed}{number of seeds randomly drawn from members of hidden population (group K)}

\item{n_coupons}{number of unique coupons given to each study participant}

\item{target_type}{one of 'sample' or 'waves'}

\item{target_n}{numeric target size of RDS sample. If \code{target_type = "sample"}, this gives maximum number of respondents to be sampled (right now the RDS network can also end before reaching sample size target). If \code{target_type = "waves"}, this gives maximum number of waves of recruitment allowed}
}
\value{

}
\description{
Sampling handler for drawing RDS sample with given characteristics from individual study population
}
\examples{
\dontrun{
sample_rds(data = get_pop_network())
}

}
