% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rds.R
\name{sample_rds}
\alias{sample_rds}
\title{Draw RDS sample from single study}
\usage{
sample_rds(
  data,
  assignment_variable = "rds",
  n_seed_hidden = 10,
  n_seed_other = 0,
  n_coupons = 3,
  target_size_rds = 80
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{assignment_variable}{character string that is used as prefix for all variables generated by RDS sampling (sample identifier, recruiter ID, wave, time of show-up)}

\item{n_seed_hidden}{number of seeds randomly drawn from members of hidden population (group K)}

\item{n_seed_other}{number of seeds randomly drawn from non-members of hidden population}

\item{n_coupons}{number of unique coupons given to each study participant}

\item{target_size_rds}{target size of RDS sample (this might need to be replaced by target number of waves)}
}
\value{

}
\description{
Sampling handler for drawing RDS sample with given characteristics from individual study population
}
\examples{
\dontrun{
sample_rds(data = get_pop_network())
}

}
