% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_sample_tls.R
\name{sample_tls}
\alias{sample_tls}
\title{Draw time-loaction (TLS) sample from single study}
\usage{
sample_tls(
  data,
  sampling_variable = "tls",
  drop_nonsampled = FALSE,
  target_n_tls = 2,
  loc_pattern = "^loc_[0-9]$"
)
}
\arguments{
\item{data}{pass-through population data frame}

\item{sampling_variable}{character string that is used as prefix for all variables generated by TLS sampling. Default is 'tls'}

\item{drop_nonsampled}{logical indicating whether to drop units that are not sampled. Default is \code{FALSE}}

\item{target_n_tls}{target number of sampled locations}

\item{loc_pattern}{character string containing regular expression to match all locality names in the study population dataset}
}
\value{
Population or sample data frame for single study with TLS sample characteristics added
}
\description{
Sampling handler for drawing TLS sample with given characteristics from individual study population
}
