% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_network.R
\name{sim_block_network}
\alias{sim_block_network}
\title{Handler for Simulation of Network Using Block Model}
\usage{
sim_block_network(
  N,
  K,
  prev_K,
  rho_K,
  p_edge_within,
  p_edge_between,
  directed = FALSE
)
}
\arguments{
\item{N}{number of units in population}

\item{K}{number of groups}

\item{prev_K}{named numeric vector of prevalence for each group with last group being hidden}

\item{rho_K}{numeric vector of correlations in group memberships}

\item{p_edge_within}{named list of numeric vectors giving probability of link between in-group members and out-group members for each of groups. The order of objects in list have to follow the order of \code{prev_K}}

\item{p_edge_between}{named list of numeric values giving probability of link between in- and out-group member for each of groups. The order of objects in list have to follow the order of \code{prev_K}}

\item{directed}{logical, whether links are directed or undirected}
}
\value{
igraph network object with vertex attribute \code{type} in the binary coded format (consists of 0's and 1's only)
}
\description{
Draw a Simulation of Network Using Block Matrix of Probabilities of Edges Based on Group Memberships
}
\examples{
\dontrun{
 sim_block_network(N = 2000, K = 3,
                   prev_K = c(frame = .7,
                              known = .2,
                              hidden = .05),
                   rho_K = c(.05, .01, .01),
                   p_edge_within =
                     list(frame = c(.3,.3),
                          known = c(.3,.3),
                          hidden = c(.3,.9)),
                   p_edge_between =
                     list(frame = 0.3,
                          known = 0.3,
                          hidden = 0.05),
                   directed = FALSE)
}

}
